Correct bickfords.css build file errors
The bickfords.css file had the header and footer data added to it as
part of the build.  File has been removed from the build process and now
permanently lives in the "build" file.  Project for another day - figure
out how to include css files as part of the build process.